feat: move tasks/[...id] to app dir
