+++ .env.example
@@ -9,3 +9,9 @@ SHADOW_DATABASE_URL=
 NEXTAUTH_URL=http://localhost:3000
 NEXT_PUBLIC_REALTIME_URL=https://rtss.proginth.app/stream
 NEXT_PUBLIC_AWS_URL="https://proginth.sgp1.cdn.digitaloceanspaces.com"
+
+BUCKET_NAME=
+BUCKET_KEY_ID=
+BUCKET_KEY_SECRET=
+BUCKET_ENDPOINT=
+BUCKET_REGION=
