+++ src/components/Task/Tabs/Submit.tsx
@@ -1,3 +1,6 @@
+'use client'
+// TODO Get session on server?
+
 import { Task } from '@prisma/client'
 import { useSession } from 'next-auth/react'
